# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: Error Message
        type: string
      status:
        example: HTTP Status Error
        type: string
      status_code:
        example: 400
        type: integer
      success:
        example: false
        type: boolean
    type: object
...
  models.ResponseAny:
    properties:
      data: {}
      message:
        example: Succes Get Comment by Id Post
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@socialmedia.com
    name: API Support
  description: API documentation for social media app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media API
  version: "1.0"
paths:
...
  /users/{id}/follow:
    post:
      description: Follow another user by ID
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Success Followed
          schema:
            $ref: '#/definitions/models.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow user
      tags:
      - User
  /users/{id}/unfollow:
    delete:
      description: Unfollow another user by ID
      parameters:
      - description: Target User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unfollow user
      tags:
      - User
  /users/followers:
    get:
      description: Get list of users who follow me
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get followers
      tags:
      - User
  /users/following:
    get:
      description: Get list of users I follow
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get following
      tags:
      - User
  /users/profile:
    get:
      description: Get the profile of the logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - User
    put:
      consumes:
      - multipart/form-data
      description: Update profile fields (fullname, phone, and profile picture)
      parameters:
      - description: Full Name
        in: formData
        name: fullname
        type: string
      - description: Phone Number
        in: formData
        name: phone
        type: string
      - description: Profile Image
        in: formData
        name: img
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - User
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"